emmmm

+++ os/lib/string.c
@@ -1,6 +1,6 @@
 #include "../include/type.h"
 
-u32 strlen(char * str){
+size_t strlen(char * str){
     u32 len = 0;
     while (str[len] != 0){
         len++;
